code updated

+++ api/controllers/post.controller.js
@@ -76,6 +76,9 @@ export const getPosts = async (req, res, next) => {
 }
 
 export const deletePosts = async (req, res, next) => {
+    if(!req.user.isAdmin || req.user.id !== req.params.userId){
+        return next(errorHandler(403, "You are not allowed to delete this posts"))
+    }
 
     try {
         await Post.findByIdAndDelete(req.params.postId);
